# packages/backend/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Marvel Collection
  description: An API for trading marvel comics amongst users.
  version: 1.0.0
servers:
  - url: https://marvel-collector-backend.onrender.com/api/v1
    description: Main (production) server
  - url: http://localhost:5000/api/v1
    description: Internal development server for local testing
...
paths:
...
  /trade-offer:
    post:
      tags:
        - trade offers
      operationId: createTradeOffers
      summary: Create trade offers
      description: Create trade offers
      requestBody:
        description: Create trade offers
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                  description: type of trade offer
                  example: SELL/EXCHANGE
                comicId:
                  type: number
                  description: comic id
                  example: 1
                phoneNumber:
                  type: string
                  description: user phone number
                  example: +1-1212-1212-2323
                email:
                  type: string
                  description: user email
                  format: email
                  example: example@gmail.com
                message:
                  type: string
                  description: message
                  example: I want batman comics in return
                wantedComicId:
                  type: number
                  description: comic the user wants in return
                  example: 2
      responses:
        '201':
          description: A successful Operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    description: success
                    example: success
                  data:
                    type: object
                    description: data
                    properties:
                      tradeOfferId:
                        type: string
                        description: trade offer id
                        example: 10
                      type:
                        type: string
                        description: trade offer type
                        example: SELL/EXCHANGE
                      status:
                        type: string
                        description: trade offer status
                        example: PENDING
                      message:
                        type: string
                        description: message
                        example: I want batman comics in return
                      price:
                        type: number
                        description: comic price
                        example: 40
                      wantedComicId:
                        type: number
                        description: comic Id
                        example: 2
                      createdAt:
                        type: string
                        description: time at which the trade offer was created
                        example: 2023-04-11T17:26:14.462Z
                      createdBy:
                        type: object
                        description: user
                        properties:
                          userId:
                            type: string
                            description: user id
                            example: 1245wirito4606063jkkfjfjf
                          firstName:
                            type: string
                            description: user first name
                            example: Pearl
                          lastName:
                            type: string
                            description: user last name
                            example: Obi
                          username:
                            type: string
                            description: user name
                            example: pearlobi
                          profileImage:
                            type: string
                            description: user profile image
                            example: www.image.com/profileimage
                          location:
                            type: string
                            description: user location
                            example: Lagon State
                          contactDetails:
                            type: object
                            description: contact details
                            properties:
                              email:
                                type: string
                                description: email
                                example: example@gmail.com
                              phoneNumber:
                                type: string
                                description: user phone number
                                example: +1-1212-1212-2323
                          tradeOffer:
                            type: object
                            description: comics used in trade offer
                            properties:
                              collectionId:
                                type: string
                                description: collection Id
                                example: 121h21h2k12k12h12h12
                              comicId:
                                type: number
                                description: comic Id
                                example: 1
                              title:
                                type: string
                                description: comic title
                                example: Batman Origins
                              imageUrl:
                                type: string
                                description: comic images
                                example: www.image.com/batmanorigins
                              tradeOfferId:
                                type: string
                                description: trade offer Id
                                example: 12121gjghkkk212128y821
...
  /trade-request:
    post:
      tags:
        - trade requests
      operationId: tradeRequest
      summary: trade request
      description: Make a trade request
      requestBody:
        description: Make a trade request
        content:
          application/json:
            schema:
              type: object
              properties:
                tradeOfferId:
                  type: string
                  description: Trade offer id
                  example: 53d26fcf-234d-4555-b0d5-a5dfc91caa7e
                receiverComicId:
                  type: number
                  description: Comic id
                  example: 2
      responses:
        '201':
          description: A Successful Operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    description: success
                    example: success
                  data:
                    type: object
                    description: successfully creating a trade request
                    properties:
                      trade-request:
                        type: object
                        description: trade request
                        properties:
                          id:
                            type: string
                            description: trade request Id
                            example: d4cd859d-e090-4638-9efd-353fa986a8f6
                          receiverId:
                            type: string
                            description: user Id
                            example: f54886b0-cba4-438f-a3c2-425dc438182d
                          tradeOfferId:
                            type: string
                            description: trade offer Id
                            example: 53d26fcf-234d-4555-b0d5-a5dfc91caa7e
                          receiverComicId:
                            type: number
                            description: comic Id
                            example: 2
                          status:
                            type: string
                            description: trade request status
                            example: PENDING
                          createdAt:
                            type: string
                            description: created at time
                            format: datetime
                            example: 2023-04-13T09:05:39.946Z
      security:
        - cookieAuth: []
...
components:
...
  securitySchemes:
    cookieAuth: # arbitrary name for the security scheme; will be used in the "security" key later
      type: apiKey
      in: cookie
      name: JSESSIONID